requirements.txt
pytest==8.0.0
selenium==4.16.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
loguru==0.7.2
requests==2.31.0
fake-useragent==1.4.0
webdriver-manager==4.0.1
pytest-cov==4.1.0 